Refactor: adjusting location of ROUTE validations

+++ src/main/java/com/electricitybill/service/impl/RotaServiceImpl.java
@@ -3,10 +3,12 @@ package com.electricitybill.service.impl;
 import com.electricitybill.entity.RotaEntity;
 import com.electricitybill.repository.RotaRepository;
 import com.electricitybill.service.RotaService;
+import jakarta.persistence.NoResultException;
 import lombok.AllArgsConstructor;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Objects;
 
 @AllArgsConstructor
 public class RotaServiceImpl implements RotaService {
@@ -15,11 +17,17 @@ public class RotaServiceImpl implements RotaService {
 
     @Override
     public List<RotaEntity> findAll() {
+        if (rotaRepository.findAll().isEmpty()) {
+            throw new NoResultException("Search not found!");
+        }
         return rotaRepository.findAll();
     }
 
     @Override
     public RotaEntity findById(int id) {
+        if (Objects.isNull(rotaRepository.findById(id))) {
+            throw new NoResultException("Individual search not found!");
+        }
         return rotaRepository.findById(id);
     }
 
@@ -30,11 +38,17 @@ public class RotaServiceImpl implements RotaService {
 
     @Override
     public boolean update(int id, RotaEntity entity) throws SQLException {
+        if (Objects.isNull(rotaRepository.findById(id))) {
+            throw new NoResultException("Route not found!");
+        }
         return rotaRepository.update(id, entity);
     }
 
     @Override
     public boolean delete(int id) throws SQLException {
+        if (Objects.isNull(rotaRepository.findById(id))) {
+            throw new NoResultException("Route not found!");
+        }
         return rotaRepository.delete(id);
     }
 }
